progress on core refactoring
