+++ libs/common/src/statistics/index.ts
@@ -1,8 +1,8 @@
-export * from './common.enum';
-export * from './common.interface';
 export * from './datamart.service';
-export * from './match-service-helper';
+export * from './match-service-helper.util';
 export * from './metrics';
+export * from './statistics.enum';
+export * from './statistics.interface';
 export * from './stats.service';
 export * from './stats.util';
 export * from './time-dimension';
